✨ [채용공고 등록] JobPostForm 컴포넌트에 입력 필드 추가 및 구조 개선
