+++ src/features/recruit/components/inputs/DeadlineInput.tsx
@@ -1,0 +1,18 @@
+import { FormField } from "@/features/recruit/components/inputs";
+import { INPUT_CLASS } from "@/features/recruit/constants/classNames";
+import { JobPostFormValues } from "@/features/recruit/schemas/jobPostSchema";
+import { FieldError, UseFormRegister } from "react-hook-form";
+
+export function DeadlineInput({
+  register,
+  error,
+}: {
+  register: UseFormRegister<JobPostFormValues>;
+  error?: FieldError;
+}) {
+  return (
+    <FormField label="공고 마감일" error={error}>
+      <input type="date" {...register("deadline")} className={`w-full ${INPUT_CLASS}`} />
+    </FormField>
+  );
+}
